fix(lint): clear alerts/ + integration/ findings
- alerts/evaluator.go: build the VictoriaMetrics query via NewRequestWithContext
  + client.Do instead of client.Get (noctx) — request now carries a context,
  ready for a caller-supplied ctx; behavior unchanged (same GET, 8s timeout).
- integration/servicenow.go: #nosec G101 on headerWebhookSecret — it's an HTTP
  header NAME ("X-NetOps-Webhook-Secret"), not a credential (false positive).
- integration/providers_test.go: err != ErrSignatureInvalid → errors.Is (errorlint;
  survives future error wrapping).
- integration/reconcile_test.go: drop redundant int64(seq) conversion (unconvert).

golangci-lint ./alerts/... ./integration/... now clean; tests pass.

+++ NetOps_Observability/src/backend/integration/reconcile_test.go
@@ -6,7 +6,7 @@ import (
 )
 
 func stateEv(typ EventType, ext string, seq int64) IntegrationEvent {
-	return IntegrationEvent{Type: typ, ExternalState: ext, ExternalSeq: seq, OccurredAt: time.Unix(int64(seq), 0)}
+	return IntegrationEvent{Type: typ, ExternalState: ext, ExternalSeq: seq, OccurredAt: time.Unix(seq, 0)}
 }
 
 func TestReconcile_AppliesMappedState(t *testing.T) {
